+++ CMakeLists.txt
@@ -1,23 +1,27 @@
-
-# root cmd_processing
-
 cmake_minimum_required(VERSION 3.2)
 
-set (THIS prj_game)
+set (this OtusTankGame)
 
 if (DEFINED ENV{TRAVIS_BUILD_NUMBER})
-    project(${THIS} VERSION 0.0.$ENV{TRAVIS_BUILD_NUMBER})
+    project(${this} VERSION 0.0.$ENV{TRAVIS_BUILD_NUMBER})
 else()
-    project(${THIS} 
+    project(${this} 
             VERSION 0.0.1)
 endif()
 
+set(CMAKE_CXX_STANDARD 17)
+
 add_subdirectory(src)
+
+add_executable(OTG
+    src/main.cpp
+)
+target_link_libraries(OTG
+    OtgLib
+)
 
 if (NOT DEFINED ENV{PC_00330_80000_00000_AA472})
     
-    add_subdirectory(test)
-
     set(CPACK_GENERATOR DEB)
 
     set(CPACK_PACKAGE_VERSION_MAJOR "${PROJECT_VERSION_MAJOR}")
@@ -30,3 +34,5 @@ if (NOT DEFINED ENV{PC_00330_80000_00000_AA472})
 
 endif()
 
+add_subdirectory(test)
+enable_testing()
