+++ test/CMakeLists.txt
@@ -1,52 +1,42 @@
-# test prj_game
-
 cmake_minimum_required(VERSION 3.2)
 
-set(THIS tb_game)
-set(SOURCES tb_main)
-
-if (DEFINED ENV{TRAVIS_BUILD_NUMBER})
-    project(${THIS} VERSION 0.0.$ENV{TRAVIS_BUILD_NUMBER})
-else()
-    project(${THIS} VERSION 0.0.1)
+if (DEFINED ENV{PC_00330_80000_00000_AA472})
+    set(GTEST_ROOT $ENV{GTEST_LIBRARY_DIR})
+    set(GTEST_INCLUDE_DIR $ENV{GTEST_INCLUDE_DIR})
 endif()
 
 find_package(GTest REQUIRED)
-find_package(Threads REQUIRED)
 
-add_executable(${THIS} ${SOURCES})
+set(this tb_main)
+set(source tb_main.cpp)
 
-set_target_properties(
-    ${THIS}
-    PROPERTIES
-    CXX_STANDARD 17
-    CXX_STANDARD_REQUIRED ON
-)
+add_executable(${this} ${source})
 
 target_include_directories(
-    ${THIS}
+    ${this}
     PRIVATE
-    ${CMAKE_CURRENT_SOURCE_DIR}/../lib
+    ${GTEST_INCLUDE_DIR}
 )
 
-target_link_libraries(${THIS} ${GTEST_BOTH_LIBRARIES})
-target_link_libraries(${THIS} ${CMAKE_THREAD_LIBS_INIT})
-target_link_libraries(${THIS} ${LIBS})
+target_link_libraries(${this} ${GTEST_BOTH_LIBRARIES})
 
 if (MSVC)
     target_compile_options(
-        ${THIS}
+        ${this}
         PRIVATE
         /W4
+        $<$<CONFIG:Debug>:/MT>
+        $<$<CONFIG:Release>:/MT>
+
     )
 else ()
     target_compile_options(
-        ${THIS}
+        ${this}
         PRIVATE
         -Wall -Wextra -pedantic -Werror
     )
 endif()
 
-enable_testing()
+add_test(${this} ${this})
 
-add_test(${THIS} ${THIS})
+
